verify certificate chain for broker app
Google does a fix in 4.1_r4 to return a valid certificate chain, before
the fix, we're under the threat for talking to a fake broker app.
Attacker can take the certificate from a valid broker app, and create a
fake certificate chain. Since the lower sdk doesn't perform certificate
chain validation, and we only check if the returned signatures only a
valid one.

The fix will create a certificate chain based on the returned signatures
from the os, and validate the certificate chain.