Promote string literals in base class to members
Strickly speaking, we only need to do this for the "stop_time" only.
But it is possibly OK to increase the blast radius to other literals,
given that:

1. This is the base class, so there may be a slim chance that so other
helper methods in future will need to access those fields.
2. Because I chose to use string literal in the beginning, I actually
already forgot and then duplicated "stop_time" in my 2 methods. To prevent
this from happening again, it may be understandable to go with the string
const approach.