Specify JDK path as JAVA_HOME in yml
Xamarin searches for the version of Java to use by starting at the
oldest version and working upwards, stopping when it finds one.
This causes a build failure with the Xamarin Forms sample app; set the
JavaSdkDirectory environment variable to specify the Java version as set
in the JAVA_HOME env-var on the AppVeyor agent.