Revert changes to string-formatting for Error messages
Since we don't have good enough test coverage or knowledge of what (if
anything) will be affected by changing the error message strings, we
should not change these (even if they're wrong) for now.

It could be the case that a consumer of ADAL has taken a dependency on
the exact form of the error messages returned from ADAL as part of error
handling (for example).