Handle new RSA default (RSACng) on netfx4.7
The default RSA implementation has changed in .NET Framework 4.7.
Previously we got handed back a instance of RSACryptoServiceProvider,
whereas now we get a type of RSACng.

Because we're targeting .NET Framework 4.5 and RSACng was only
introduced in .NET Framework 4.6, we cannot actually call methods on
that type (without using reflection anyway..).

Instead we now try and construct and initialise an new instance of
RSACryptoServiceProvider using the private key parameters in the X509
certificate.